+++ spec/factories/stats.rb
@@ -6,6 +6,8 @@ FactoryBot.define do
     month { 1 }
     distance { 1000 } # 1 km
     user
+    sharing_settings { {} }
+    sharing_uuid { SecureRandom.uuid }
     toponyms do
       [
         {
@@ -15,6 +17,32 @@ FactoryBot.define do
           'country' => 'Russia'
         }, { 'cities' => [], 'country' => nil }
       ]
+    end
+
+    trait :with_sharing_enabled do
+      after(:create) do |stat, evaluator|
+        stat.enable_sharing!(expiration: 'permanent')
+      end
+    end
+
+    trait :with_sharing_disabled do
+      sharing_settings do
+        {
+          'enabled' => false,
+          'expiration' => nil,
+          'expires_at' => nil
+        }
+      end
+    end
+
+    trait :with_sharing_expired do
+      sharing_settings do
+        {
+          'enabled' => true,
+          'expiration' => '1h',
+          'expires_at' => 1.hour.ago.iso8601
+        }
+      end
     end
   end
 end